requirements.txt
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
idna==3.6
imutils==0.5.4
ipykernel==6.28.0
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.6
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.5.9
numpy==2.0.1
opencv-contrib-python==4.10.0.84
packaging==23.2
pandas==2.2.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
plotly==5.23.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
tenacity==9.0.0
tkinterdnd2==0.3.0
tornado==6.4
traitlets==5.14.1
ttkbootstrap==1.10.1
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.3
